Define BadInstructionException for non-x86_64 arch
Carthage, which sets ONLY_ACTIVE_ARCH = NO, will cause the swift
compiler to generate a bridging header with a random architecture set.
This means there is non-deterministic behavior for when
BadInstructionException is defined in the bridging header. But
CwlCatchBadInstruction.m needs BadInstructionException to be defined.

BadInstructionException implementation will fatally error if used on
non-x86_64 architectures.